Device+shader cache moved to separate file
