+++ src/rhi-shared-fwd.h
@@ -1,0 +1,15 @@
+#pragma once
+
+namespace rhi {
+class Device;
+class CommandEncoder;
+class CommandList;
+class ShaderObjectLayout;
+class ShaderProgram;
+class Pipeline;
+class ShaderObject;
+class RootShaderObject;
+struct ExtendedShaderObjectTypeList;
+
+typedef uint32_t ShaderComponentID;
+} // namespace rhi
