+++ src/rhi-shared.h
@@ -9,6 +9,10 @@
 #include "core/common.h"
 #include "core/short_vector.h"
 
+#include "rhi-shared-fwd.h"
+
+#include "device.h"
+
 #include <map>
 #include <set>
 #include <memory>
@@ -294,7 +298,6 @@ public:
     StructHolder m_descHolder;
 };
 
-typedef uint32_t ShaderComponentID;
 const ShaderComponentID kInvalidComponentID = 0xFFFFFFFF;
 
 struct ExtendedShaderObjectType
@@ -1140,96 +1143,7 @@ public:
     virtual SLANG_NO_THROW Result SLANG_MCALL getNativeHandle(NativeHandle* outHandle) override;
 };
 
-struct ComponentKey
-{
-    std::string typeName;
-    short_vector<ShaderComponentID> specializationArgs;
-    size_t hash;
-    void updateHash()
-    {
-        hash = std::hash<std::string_view>()(typeName);
-        for (auto& arg : specializationArgs)
-            hash_combine(hash, arg);
-    }
-    template<typename KeyType>
-    bool operator==(const KeyType& other) const
-    {
-        if (typeName != other.typeName)
-            return false;
-        if (specializationArgs.size() != other.specializationArgs.size())
-            return false;
-        for (size_t i = 0; i < other.specializationArgs.size(); i++)
-        {
-            if (specializationArgs[i] != other.specializationArgs[i])
-                return false;
-        }
-        return true;
-    }
-};
 
-struct PipelineKey
-{
-    Pipeline* pipeline;
-    short_vector<ShaderComponentID> specializationArgs;
-    size_t hash;
-    void updateHash()
-    {
-        hash = std::hash<void*>()(pipeline);
-        for (auto& arg : specializationArgs)
-            hash_combine(hash, arg);
-    }
-    bool operator==(const PipelineKey& other) const
-    {
-        if (pipeline != other.pipeline)
-            return false;
-        if (specializationArgs.size() != other.specializationArgs.size())
-            return false;
-        for (size_t i = 0; i < other.specializationArgs.size(); i++)
-        {
-            if (specializationArgs[i] != other.specializationArgs[i])
-                return false;
-        }
-        return true;
-    }
-};
-
-// A cache from specialization keys to a specialized `ShaderKernel`.
-class ShaderCache : public RefObject
-{
-public:
-    ShaderComponentID getComponentId(slang::TypeReflection* type);
-    ShaderComponentID getComponentId(std::string_view name);
-    ShaderComponentID getComponentId(ComponentKey key);
-
-    RefPtr<Pipeline> getSpecializedPipeline(PipelineKey programKey)
-    {
-        auto it = specializedPipelines.find(programKey);
-        if (it != specializedPipelines.end())
-            return it->second;
-        return nullptr;
-    }
-
-    void addSpecializedPipeline(PipelineKey key, RefPtr<Pipeline> specializedPipeline);
-
-    void free()
-    {
-        componentIds = decltype(componentIds)();
-        specializedPipelines = decltype(specializedPipelines)();
-    }
-
-protected:
-    struct ComponentKeyHasher
-    {
-        std::size_t operator()(const ComponentKey& k) const { return k.hash; }
-    };
-    struct PipelineKeyHasher
-    {
-        std::size_t operator()(const PipelineKey& k) const { return k.hash; }
-    };
-
-    std::unordered_map<ComponentKey, ShaderComponentID, ComponentKeyHasher> componentIds;
-    std::unordered_map<PipelineKey, RefPtr<Pipeline>, PipelineKeyHasher> specializedPipelines;
-};
 
 static const int kRayGenRecordSize = 64; // D3D12_RAYTRACING_SHADER_TABLE_BYTE_ALIGNMENT;
 
@@ -1276,196 +1190,6 @@ public:
 };
 
 
-// Device implementation shared by all platforms.
-// Responsible for shader compilation, specialization and caching.
-class Device : public IDevice, public ComObject
-{
-public:
-    SLANG_COM_OBJECT_IUNKNOWN_ADD_REF
-    SLANG_COM_OBJECT_IUNKNOWN_RELEASE
-
-    virtual SLANG_NO_THROW Result SLANG_MCALL getNativeDeviceHandles(DeviceNativeHandles* outHandles) override;
-    virtual SLANG_NO_THROW Result SLANG_MCALL
-    getFeatures(const char** outFeatures, size_t bufferSize, uint32_t* outFeatureCount) override;
-    virtual SLANG_NO_THROW bool SLANG_MCALL hasFeature(const char* featureName) override;
-    virtual SLANG_NO_THROW Result SLANG_MCALL getFormatSupport(Format format, FormatSupport* outFormatSupport) override;
-    virtual SLANG_NO_THROW Result SLANG_MCALL getSlangSession(slang::ISession** outSlangSession) override;
-    virtual SLANG_NO_THROW Result SLANG_MCALL queryInterface(const SlangUUID& uuid, void** outObject) override;
-    IDevice* getInterface(const Guid& guid);
-
-    virtual SLANG_NO_THROW Result SLANG_MCALL
-    createTextureFromNativeHandle(NativeHandle handle, const TextureDesc& srcDesc, ITexture** outTexture) override;
-
-    virtual SLANG_NO_THROW Result SLANG_MCALL createTextureFromSharedHandle(
-        NativeHandle handle,
-        const TextureDesc& srcDesc,
-        const Size size,
-        ITexture** outTexture
-    ) override;
-
-    virtual SLANG_NO_THROW Result SLANG_MCALL
-    createBufferFromNativeHandle(NativeHandle handle, const BufferDesc& srcDesc, IBuffer** outBuffer) override;
-
-    virtual SLANG_NO_THROW Result SLANG_MCALL
-    createBufferFromSharedHandle(NativeHandle handle, const BufferDesc& srcDesc, IBuffer** outBuffer) override;
-
-    virtual SLANG_NO_THROW Result SLANG_MCALL
-    createInputLayout(const InputLayoutDesc& desc, IInputLayout** outLayout) override;
-
-    virtual SLANG_NO_THROW Result SLANG_MCALL
-    createRenderPipeline(const RenderPipelineDesc& desc, IRenderPipeline** outPipeline) override;
-    virtual SLANG_NO_THROW Result SLANG_MCALL
-    createComputePipeline(const ComputePipelineDesc& desc, IComputePipeline** outPipeline) override;
-    virtual SLANG_NO_THROW Result SLANG_MCALL
-    createRayTracingPipeline(const RayTracingPipelineDesc& desc, IRayTracingPipeline** outPipeline) override;
-
-    virtual SLANG_NO_THROW Result SLANG_MCALL createShaderObject(
-        slang::ISession* session,
-        slang::TypeReflection* type,
-        ShaderObjectContainerType containerType,
-        IShaderObject** outObject
-    ) override;
-
-    virtual SLANG_NO_THROW Result SLANG_MCALL
-    createShaderObjectFromTypeLayout(slang::TypeLayoutReflection* typeLayout, IShaderObject** outObject) override;
-
-    virtual SLANG_NO_THROW Result SLANG_MCALL
-    createRootShaderObject(IShaderProgram* program, IShaderObject** outObject) override;
-
-    // Provides a default implementation that returns SLANG_E_NOT_AVAILABLE for platforms
-    // without ray tracing support.
-    virtual SLANG_NO_THROW Result SLANG_MCALL getAccelerationStructureSizes(
-        const AccelerationStructureBuildDesc& desc,
-        AccelerationStructureSizes* outSizes
-    ) override;
-
-    // Provides a default implementation that returns SLANG_E_NOT_AVAILABLE for platforms
-    // without ray tracing support.
-    virtual SLANG_NO_THROW Result SLANG_MCALL createAccelerationStructure(
-        const AccelerationStructureDesc& desc,
-        IAccelerationStructure** outAccelerationStructure
-    ) override;
-
-    // Provides a default implementation that returns SLANG_E_NOT_AVAILABLE for platforms
-    // without ray tracing support.
-    virtual SLANG_NO_THROW Result SLANG_MCALL
-    createShaderTable(const ShaderTableDesc& desc, IShaderTable** outTable) override;
-
-    // Provides a default implementation that returns SLANG_E_NOT_AVAILABLE.
-    virtual SLANG_NO_THROW Result SLANG_MCALL createFence(const FenceDesc& desc, IFence** outFence) override;
-
-    // Provides a default implementation that returns SLANG_E_NOT_AVAILABLE.
-    virtual SLANG_NO_THROW Result SLANG_MCALL
-    waitForFences(uint32_t fenceCount, IFence** fences, uint64_t* fenceValues, bool waitForAll, uint64_t timeout)
-        override;
-
-    // Provides a default implementation that returns SLANG_E_NOT_AVAILABLE.
-    virtual SLANG_NO_THROW Result SLANG_MCALL
-    getTextureAllocationInfo(const TextureDesc& desc, Size* outSize, Size* outAlignment) override;
-
-    // Provides a default implementation that returns SLANG_E_NOT_AVAILABLE.
-    virtual SLANG_NO_THROW Result SLANG_MCALL getTextureRowAlignment(size_t* outAlignment) override;
-
-    // Provides a default implementation that returns SLANG_E_NOT_AVAILABLE.
-    virtual SLANG_NO_THROW Result SLANG_MCALL createSurface(WindowHandle windowHandle, ISurface** outSurface) override;
-
-    // Provides a default implementation that returns SLANG_E_NOT_AVAILABLE.
-    virtual SLANG_NO_THROW Result SLANG_MCALL
-    getCooperativeVectorProperties(CooperativeVectorProperties* properties, uint32_t* propertyCount) override;
-
-    // Provides a default implementation that returns SLANG_E_NOT_AVAILABLE.
-    virtual SLANG_NO_THROW Result SLANG_MCALL
-    convertCooperativeVectorMatrix(const ConvertCooperativeVectorMatrixDesc* descs, uint32_t descCount) override;
-
-    Result getEntryPointCodeFromShaderCache(
-        slang::IComponentType* program,
-        SlangInt entryPointIndex,
-        SlangInt targetIndex,
-        slang::IBlob** outCode,
-        slang::IBlob** outDiagnostics = nullptr
-    );
-
-    Result getShaderObjectLayout(
-        slang::ISession* session,
-        slang::TypeReflection* type,
-        ShaderObjectContainerType container,
-        ShaderObjectLayout** outLayout
-    );
-
-    Result getShaderObjectLayout(
-        slang::ISession* session,
-        slang::TypeLayoutReflection* typeLayout,
-        ShaderObjectLayout** outLayout
-    );
-
-public:
-    inline void handleMessage(DebugMessageType type, DebugMessageSource source, const char* message)
-    {
-        m_debugCallback->handleMessage(type, source, message);
-    }
-
-    inline void warning(const char* message)
-    {
-        handleMessage(DebugMessageType::Warning, DebugMessageSource::Layer, message);
-    }
-
-    Result createShaderObject(ShaderObjectLayout* layout, ShaderObject** outObject);
-    Result createRootShaderObject(ShaderProgram* program, RootShaderObject** outObject);
-
-    Result getSpecializedProgram(
-        ShaderProgram* program,
-        const ExtendedShaderObjectTypeList& specializationArgs,
-        ShaderProgram** outSpecializedProgram
-    );
-
-    Result specializeProgram(
-        ShaderProgram* program,
-        const ExtendedShaderObjectTypeList& specializationArgs,
-        ShaderProgram** outSpecializedProgram
-    );
-
-    Result getConcretePipeline(
-        Pipeline* pipeline,
-        ExtendedShaderObjectTypeList* specializationArgs,
-        Pipeline*& outPipeline
-    );
-
-    virtual Result createShaderObjectLayout(
-        slang::ISession* session,
-        slang::TypeLayoutReflection* typeLayout,
-        ShaderObjectLayout** outLayout
-    ) = 0;
-
-    virtual Result createRootShaderObjectLayout(
-        slang::IComponentType* program,
-        slang::ProgramLayout* programLayout,
-        ShaderObjectLayout** outLayout
-    ) = 0;
-
-    virtual void customizeShaderObject(ShaderObject* shaderObject) { SLANG_UNUSED(shaderObject); }
-
-    virtual Result createRenderPipeline2(const RenderPipelineDesc& desc, IRenderPipeline** outPipeline);
-    virtual Result createComputePipeline2(const ComputePipelineDesc& desc, IComputePipeline** outPipeline);
-    virtual Result createRayTracingPipeline2(const RayTracingPipelineDesc& desc, IRayTracingPipeline** outPipeline);
-
-protected:
-    virtual SLANG_NO_THROW Result SLANG_MCALL initialize(const DeviceDesc& desc);
-
-protected:
-    std::vector<std::string> m_features;
-    std::vector<CooperativeVectorProperties> m_cooperativeVectorProperties;
-
-public:
-    SlangContext m_slangContext;
-    ShaderCache m_shaderCache;
-
-    ComPtr<IPersistentShaderCache> m_persistentShaderCache;
-
-    std::map<slang::TypeLayoutReflection*, RefPtr<ShaderObjectLayout>> m_shaderObjectLayoutCache;
-    ComPtr<IPipelineCreationAPIDispatcher> m_pipelineCreationAPIDispatcher;
-
-    IDebugCallback* m_debugCallback = nullptr;
-};
 
 bool isDepthFormat(Format format);
 bool isStencilFormat(Format format);
